refactor: work entity
